pyproject.toml
[project]
name = "vibepascal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=1.63.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "fastapi>=0.133.1",
    "uvicorn>=0.41.0",
    "pytest>=8.0.0",
    "strands-agents>=1.28.0",
    "litellm>=1.81.16",
    "agent-framework>=1.0.0rc3",
    "agno>=2.5.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
